requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
apispec==6.3.0
apispec-webframeworks==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
flask-apispec==0.11.4
Flask-Cors==3.0.10
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow2==2.20.2
multidict==6.0.4
packaging==23.0
passlib==1.7.4
protobuf==4.21.12
pycparser==2.21
pycryptodomex==3.16.0
pyseto==1.7.0
PyYAML==6.0
replit==3.2.5
requests==2.28.2
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.14
webargs==8.2.0
Werkzeug==2.2.2
yarl==1.8.2
